requirements.txt
paramiko
psutil
python-telegram-bot
pytelegrambotapi
mysql-connector-python